implement a modal wrapper that takes in a component for its body
